// src/app/common/models/90contents.models.ts
export class Contents {
    id?: Number;
    typeid?: string;
    siteid?: string;
    title?: string;
    url?: string;
    summary?: string;
    content?: string;
    pathfile?: string;
    pathimage?: string;
    comment?: string;
    postdate?: any;
    changedate?: string;
    approved?: any;
    lang?: string;
    author?: string;
}